+++ Web/Jobs/ChuKuJob.cs
@@ -61,7 +61,7 @@ namespace Web.Jobs
              {
                  WarehouseTraySpecification warehouseTraySpecification = new WarehouseTraySpecification(null,
                      null,
-                     new List<int> {0, 0}, null, null, null,
+                     new List<double> {0, 0}, null, null, null,
                      new List<int> {Convert.ToInt32(TRAY_STEP.待出库)}, null,
                      null, null, null, null);
